requirements.txt
python-dotenv>=0.19.0
openai>=1.0.0
pytest>=7.0.0